fix(ci): deduplicate Mac OS build script

+++ scripts/mac-release.sh
@@ -23,45 +23,73 @@ fi
 
 COMMIT=$(git rev-parse HEAD)
 
+RELEASE_TAG=""
+# the release commit is assigned multiple tags. Only use the tag from the git run.
+if [ "${GITHUB_EVENT_NAME}" = "release" ] && [ -n "${GITHUB_REF_NAME}" ]; then
+  RELEASE_TAG="${GITHUB_REF_NAME}"
+fi
+
 os=$(uname)
 arch=$(uname -m)
 os_and_arch=${os}-${arch}
 
 # cspell:words czvf
 function tar_binary {
+  if [ "${DRY_RUN}" = "true" ]; then
+    echo "DRY RUN MODE: Would run 'tar -C $1 -czvf $1.tar.gz ${os_and_arch}'"
+    touch $1.tar.gz
+    return
+  fi
   mkdir -p $1/${os_and_arch}
   cp target/release/$1 $1/${os_and_arch}/
   tar -C $1 -czvf $1.tar.gz ${os_and_arch}
 }
 
-make $release
+if [ "${DRY_RUN}" = "true" ]; then
+  echo "DRY RUN MODE: Would run 'make $release'"
+else
+  make $release
+fi
 
 function upload_binary {
-	if [ "$release" == "release" ]
+	DRY_RUN_FLAG=""
+    if [ "${DRY_RUN}" = "true" ]; then
+        DRY_RUN_FLAG="--dryrun"
+    fi
+    if [ "$release" == "release" ]
 	then
 		tar_binary $1
 		tar_file=$1.tar.gz
 
-		if [ "$arch" == "x86_64" ]
-		then
-			aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os}/${BRANCH}/$1
-			aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os}/${BRANCH}/${COMMIT}/$1
-			aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os}/${BRANCH}/${COMMIT}/stable/$1
+	    upload_targets=()
+		if [ -n "${BRANCH}" ]; then
+			upload_targets+=("${BRANCH}")
+		fi
+		if [ -n "${RELEASE_TAG}" ] && [ "${BRANCH}" != "${RELEASE_TAG}" ]; then
+			upload_targets+=("${RELEASE_TAG}")
 		fi
 
-		aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/$1
-		aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/$1
-		aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/stable/$1
+		if [ ${#upload_targets[@]} -eq 0 ]; then
+			echo "Unable to determine upload target for release artifacts" >&2
+			exit 1
+		fi
 
-		aws s3 cp --acl public-read ${tar_file} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${tar_file}
-		aws s3 cp --acl public-read ${tar_file} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/${tar_file}
-		aws s3 cp --acl public-read ${tar_file} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/stable/${tar_file}
+		sources=("target/release/$1" "${tar_file}")
+        destinations=("$1" "${tar_file}")
+
+		for target in "${upload_targets[@]}"; do  
+			for i in "${!sources[@]}"; do
+				src=${sources[$i]}
+				dst=${destinations[$i]}
+				
+				aws s3 cp --acl public-read ${src} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${dst} ${DRY_RUN_FLAG}
+				aws s3 cp --acl public-read ${src} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/${dst} ${DRY_RUN_FLAG}
+				aws s3 cp --acl public-read ${src} s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/stable/${dst} ${DRY_RUN_FLAG}
+			done
+		done
 	else
-		if [ "$arch" == "x86_64" ]
-		then
-			aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os}/${BRANCH}/${COMMIT}/nightly/$1
-		fi
-		aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/nightly/$1
+	    folder="${release%-release}"
+		aws s3 cp --acl public-read target/release/$1 s3://build.nearprotocol.com/nearcore/${os_and_arch}/${BRANCH}/${COMMIT}/${folder}/$1 ${DRY_RUN_FLAG}
 	fi
 }
